+++ src/Henry.Api/Data/Appointment.cs
@@ -3,12 +3,12 @@
     public class Appointment
     {
         public int Id { get; set; }
-        public required Provider Provider { get; set; }
-        public Client Client { get; set; }
-        public DateOnly AppointmentOn { get; set; }
-        public TimeOnly AppointmentFrom { get; set; }
-        public TimeOnly AppointmentTo { get; set; }
-        public DateTime ReservedOn { get; set; }
-        public  bool Confirmed { get; set; }
+        public required string ProviderName { get; set; }
+        public string? ClientName { get; set; }
+        public DateOnly? AppointmentOn { get; set; }
+        public TimeOnly? AppointmentFrom { get; set; }
+        public TimeOnly? AppointmentTo { get; set; }
+        public DateTime? ReservedOn { get; set; }
+        public bool Confirmed { get; set; }
     }
 }