+++ src/Henry.Api/Data/Client.cs
@@ -1,9 +1,0 @@
-﻿namespace Henry.Api.Data
-{
-    public class Client
-    {
-        public int Id { get; set; }
-        public string? Name { get; set; }
-        public string? Email { get; set; }
-    }
-}