+++ src/Henry.Api/Data/Provider.cs
@@ -1,8 +1,0 @@
-﻿namespace Henry.Api.Data
-{
-    public class Provider
-    {
-        public int Id { get; set; }
-        public string? Name { get; set; }
-    }
-}