+++ src/Henry.Api/Migrations/20231217091718_InitialCreate.cs
@@ -1,0 +1,41 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace Henry.Api.Migrations
+{
+    /// <inheritdoc />
+    public partial class InitialCreate : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.CreateTable(
+                name: "Appointments",
+                columns: table => new
+                {
+                    Id = table.Column<int>(type: "INTEGER", nullable: false)
+                        .Annotation("Sqlite:Autoincrement", true),
+                    ProviderName = table.Column<string>(type: "TEXT", nullable: false),
+                    ClientName = table.Column<string>(type: "TEXT", nullable: true),
+                    AppointmentOn = table.Column<DateOnly>(type: "TEXT", nullable: true),
+                    AppointmentFrom = table.Column<TimeOnly>(type: "TEXT", nullable: true),
+                    AppointmentTo = table.Column<TimeOnly>(type: "TEXT", nullable: true),
+                    ReservedOn = table.Column<DateTime>(type: "TEXT", nullable: true),
+                    Confirmed = table.Column<bool>(type: "INTEGER", nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_Appointments", x => x.Id);
+                });
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropTable(
+                name: "Appointments");
+        }
+    }
+}