+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0 - 2023-12-16]
+## [0.3.0] - 2023-12-17
+
+## Changed
+
+- Refactored data down to a single table for simplicity
+- Refactored api down to just appointments for simplicity
+
+## [0.2.0] - 2023-12-16
 
 ## Added